Novena versión (falta modificar equipos y jugadores)

+++ futbol/src/main/java/com/pedro/futbol/repositorios/EquipoRepositorio.java
@@ -10,7 +10,7 @@ import com.pedro.futbol.entidades.Equipo;
 
 @Repository
 public interface EquipoRepositorio extends JpaRepository<Equipo, Long>{
-	Equipo findById(int id);
+	Equipo findById(long id);
 	
 	Equipo findByNombre(String nombre);
